face_recog_android/requirements.txt
Django==2.2.7
django-cors-headers==3.1.1
django-heroku==0.3.1
djangorestframework==3.10.3
dlib==19.18.0
face-recognition==1.2.3
gunicorn==20.0.4
imutils==0.5.2
matplotlib==3.1.1
numpy==1.16.4
opencv-contrib-python==4.1.0.25
Pillow==6.1.0
psycopg2==2.8.4
PyPDF2==1.26.0
pytz==2019.3
reportlab==3.5.23
scikit-image==0.15.0
scipy==1.3.0
whitenoise==5.0.1

